+++ test/content/simple-meta-data.meta.js
@@ -1,0 +1,13 @@
+export default {
+  title: "My most minimal SSG",
+  dateCreated: "2025-05-01 01:02:03",
+  tags: [
+    'simple',
+    'ssg',
+    'web',
+    'static',
+    'useThePlatform',
+    'noFrameworks',
+    'lightweight',
+  ]
+};