+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # V2.0.0
 - [x] make `npx run @wolframkriesing/picossg -c content -o out` work in other projects
 - [x] allow loading njk filters
+- [x] provide filter `|md` and `|mdinline`
 
 # Done
 - [x] serve and process njk files