+++ CHANGELOG.md
@@ -3,12 +3,14 @@
 - [x] make `npx run @wolframkriesing/picossg -c content -o out` work in other projects
 - [x] allow loading njk filters
 - [x] provide filter `|md` and `|mdinline`
-- [ ] make a frontmatter block work
+- [x] make a frontmatter block work
   - [x] pass meta data to the template
   - [x] make "extends/layout" key in frontmatter block work
   - [x] allow "collecting" all metadata e.g. through picossg.pages("/blog") or so and iterate over them to generate a blog list or alike
     - [x] and provide attributes for each page such as "url", ????
   - [x] pass the data into the layout template too, so e.g. a `title` of a page will be seen in the _base.njk template
+- [x] add _config.js/ts to allow pre-processing 
+- [ ] and post-processing
 
 # Done