+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# V3.0.0
+
+- [ ] rebuild tests to be at least somehow grouped/sorted
+- [ ] rewrite the processing to
+  - [ ] allow `preprocess()` and `postprocess()` functions that receive ALL files incl. data and content
+        like so: `preprocess(files: Map<filename, {content: string, data: object}>)`
+        this allows to process any file before and after the main processing, it provides access to all files e.g. to find related ones and do searches
+- [ ] allow post-processing, e.g. to minify pages
+
 # V2.0.0
 
 - [x] make `npx run @wolframkriesing/picossg -c content -o out` work in other projects
@@ -10,7 +19,6 @@
     - [x] and provide attributes for each page such as "url", ????
   - [x] pass the data into the layout template too, so e.g. a `title` of a page will be seen in the _base.njk template
 - [x] add _config.js/ts to allow pre-processing 
-- [ ] and post-processing
 
 # Done